# PERF: improve construct_1d_object_array_from_listlike