@jorisvandenbossche 2024-12-02
> Some mypy errors,

If I want something that is both Iterable and Sized, then that's Collection or Sequence ?